+++ __tests__/services/grpc/world_engine_client_grpc.service.spec.ts
@@ -15,6 +15,7 @@ import { CLIENT_ID } from '../../../src/common/constants';
 import { Logger } from '../../../src/common/logger';
 import { WorldEngineClientGrpcService } from '../../../src/services/gprc/world_engine_client_grpc.service';
 import {
+  capabilities,
   createAgent,
   extension,
   KEY,
@@ -115,7 +116,6 @@ describe('load scene', () => {
       name: SCENE,
       capabilities,
       sessionToken,
-      user,
     });
 
     const loadedAgents = result.getAgentsList();
@@ -170,6 +170,48 @@ describe('load scene', () => {
 
     expect(mockLoadScene.mock.calls[0][0].getSessionContinuation()).toEqual(
       sessionContinuation,
+    );
+  });
+
+  test('should use provided provided user name', async () => {
+    const loadScene = jest
+      .spyOn(WorldEngineClient.prototype, 'loadScene')
+      .mockImplementationOnce(mockLoadScene);
+
+    await client.loadScene({
+      name: SCENE,
+      capabilities,
+      sessionToken,
+      user: { fullName: user.fullName },
+    });
+
+    expect(loadScene).toHaveBeenCalledTimes(1);
+    expect(loadScene.mock.calls[0][0].getUser().getName()).toEqual(
+      user.fullName,
+    );
+  });
+
+  test('should use provided provided user profile', async () => {
+    const loadScene = jest
+      .spyOn(WorldEngineClient.prototype, 'loadScene')
+      .mockImplementationOnce(mockLoadScene);
+
+    await client.loadScene({
+      name: SCENE,
+      capabilities,
+      sessionToken,
+      user: { profile: user.profile },
+    });
+
+    const profileFields = loadScene.mock.calls[0][0]
+      .getUserSettings()
+      .getPlayerProfile()
+      .getFieldsList();
+
+    expect(loadScene).toHaveBeenCalledTimes(1);
+    expect(profileFields[0].getFieldId()).toEqual(user.profile.fields[0].id);
+    expect(profileFields[0].getFieldValue()).toEqual(
+      user.profile.fields[0].value,
     );
   });
 });