+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Allow to propagate user profile fields
+
 ## 2023-07-06 v1.5.1
 * Added JS examples
 * Added Say hello example